fix: removed large spacing between elements in rendered markdown (#98)
This pull request introduces changes to improve markdown formatting and
simplify the handling of content in various components. The most notable
updates include the addition of a system prompt for consistent markdown
rendering, adjustments to `MarkdownContent` for spacing between
elements, and the removal of redundant wrapper `div` elements around
markdown content.

Closes #97

**Before**

![CleanShot 2025-07-04 at 17 58
29@2x](https://github.com/user-attachments/assets/186e8234-197f-409d-8019-712437ec7043)

**After**

![CleanShot 2025-07-04 at 17 57
19@2x](https://github.com/user-attachments/assets/caa60d1f-7139-482b-99b1-51525f653422)